+++ packages/flutter_robusta/lib/flutter_robusta.dart
@@ -3,6 +3,7 @@ library flutter_robusta;
 
 export 'package:robusta_runner/robusta_runner.dart';
 
+export 'src/app_extension.dart';
 export 'src/cupertino.dart';
 export 'src/material.dart';
 export 'src/router.dart';