+++ packages/flutter_robusta/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+- BC: replace `MaterialExtension` and `CupertinoExtension` with `AppExtension`.
+- Support settings `AppExtension` via `Configurator` extension.
+
 ## 0.1.0+1
 - Add README doc